Adding the ability to collapse each folder in a result.
I started using slskd from my phone and the amount of scrolling I have to do is a lot, especially when I know the result is not the match I want based on the title of the folder. So I added the ability to toggle the state of the folders in both the transfer lists and the more generic FileList.