Add 'get full directory' to FileList
This commit adds a new {username}/directory/{directory} endpoint which
returns the contents of a given directory from a user. This has been
implemented in the frontend by adding a "Get Full Directory" button to
FileList, which is rendered if a "getFullDirectory" function is passed a
as a prop. This is not needed in Browse, so this is currently just added
to the search.