Remember theme setting between page loads
Previously, the theme setting would default to the device preference on
each load. So, if a user preferred light theme for slskd despite having
a dark device theme, they would need to toggle the theme on each page
load. By saving the theme preference in localStorage, we can override
the device default with the user's manual preference.